Implement a switch to toggle the visibility feature.
When the feature is disabled by this switch users
cannot change the visibility of their projects/repos.
If there are already private projects when the feature
gets disabled, those projects will not stay too private.

The idea is not to make everything public when the feature
is switched off, but to not execute all the (horrible)
delete_ifs etc on events etc because it's slow when there
are a lot of events like in gitorious.org. Of course it'd
be better to get rid of the delete_ifs by doing different
queries depending on whether the feature is enabled or not.
That would make it possible to implement the switch in a
much better way too.

Signed-off-by: Tero Hänninen <tero.j.hanninen@jyu.fi>